convert quotations into the ast
This requires choosing a representation for values.
Currently, this representation will be
- strings are represented as strings
- numbers are represented as numbers
- atoms and lists are represented as $-tagged objects
This should probably be better formalized at some
point, with pretty-printing adjusted to match.
Strings and numbers should also be $-tagged for
consistency.

Use debug_repr in pretty_print to print objecty values
better.

Github-issue: #5 (not a resolution)