cps: add desired name to return continuation
To reduce unnessary bindings, track the name return
continuations will be bound to, and use it when
possible. If this is possible, then a corresponding
let binding may be removed from the CPS form.

An example of an unnessary binding is:

    (klambda (x)
      (let ((a x)) ; <-- this binding here
        ...))

It's instead better to do

    (klambda (a)
      ...)

to prevent unneeded nodes and to simplify data flow.